deleted xml testfilewritertests
