+++ src/Syringe.Tests/Integration/Core/Tests/Repositories/Json/JsonExamples/50-cases.xml
