+++ src/Syringe.Tests/Integration/Core/Tests/Repositories/Json/JsonExamples/all-attributes.xml
