+++ src/Syringe.Tests/Integration/Core/Tests/Repositories/Json/JsonExamples/assertions.xml
