+++ src/Syringe.Tests/Integration/Core/Tests/Repositories/Json/JsonExamples/beforeExecuteScript.xml
