+++ src/Syringe.Tests/Integration/Core/Tests/Repositories/Json/JsonExamples/capturedvariables.json
@@ -1,0 +1,37 @@
+﻿{
+  "Tests": {
+    "variables": {
+      "variable": {
+        "-name": "baseurl",
+        "#text": "http://demo.roadkillwiki.net"
+      }
+    },
+    "test": [
+      {
+        "-shortdescription": "Roadkill",
+        "-url": "{baseurl}",
+        "-method": "get",
+        "-expectedhttpstatuscode": "200",
+        "capturedvariables": {
+          "variable": {
+            "-description": "1",
+            "#cdata-section": "This site is hosted on (.*?)<"
+          }
+        }
+      },
+      {
+        "-shortdescription": "Go to new website",
+        "-url": "http://www.{parsedresponse1}.com",
+        "-method": "get",
+        "-expectedhttpstatuscode": "200",
+        "assertions": {
+          "verify": {
+            "-type": "positive",
+            "-description": "Page title",
+            "#text": "Microsoft Azure"
+          }
+        }
+      }
+    ]
+  }
+}
