+++ src/Syringe.Tests/Integration/Core/Tests/Repositories/Json/JsonExamples/capturedvariables.xml
