+++ src/Syringe.Tests/Integration/Core/Tests/Repositories/Json/JsonExamples/large-file.xml
